requirements.txt
ultralytics
streamlit
opencv-python-headless
numpy
Pillow
torch
torchvision
supervision
